+++ src/main/java/com/programmers/voucher/strategy/VoucherValidationStrategy.java
@@ -1,5 +1,5 @@
 package com.programmers.voucher.strategy;
 
 public interface VoucherValidationStrategy {
-    void validateAmount(int amount);
+    void validateAmount(long amount);
 }
