+++ src/test/java/com/programmers/voucher/testutil/VoucherTestUtil.java
@@ -7,11 +7,11 @@ import com.programmers.voucher.domain.Voucher;
 import java.util.UUID;
 
 public class VoucherTestUtil {
-    public static Voucher createFixedVoucher(UUID voucherId, int amount) {
+    public static Voucher createFixedVoucher(UUID voucherId, long amount) {
         return new FixedAmountVoucher(voucherId, amount);
     }
 
-    public static Voucher createPercentVoucher(UUID voucherId, int amount) {
+    public static Voucher createPercentVoucher(UUID voucherId, long amount) {
         return new PercentDiscountVoucher(voucherId, amount);
     }
 }
